requirements.txt
Flask>=3.0,<4
paho-mqtt>=1.6,<3
